refactor: align provider and test suite boundaries
