+++ lib/src/features/stories/providers/stories_by_author.dart
@@ -1,6 +1,6 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:spark/src/core/network/atproto/data/models/feed_models.dart';
-import 'package:spark/src/features/stories/providers/story_provider_dependencies.dart';
+import 'package:spark/src/features/stories/providers/story_repository_provider.dart';
 import 'package:sprk_poptart/so/sprk/actor/defs.dart';
 
 part 'stories_by_author.g.dart';
