+++ lib/src/features/stories/providers/story_manager_provider.dart
@@ -1,12 +1,20 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:get_it/get_it.dart';
 import 'package:poptart/poptart.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:spark/src/core/network/atproto/data/models/feed_models.dart';
 import 'package:spark/src/core/network/atproto/data/repositories/story_repository.dart';
+import 'package:spark/src/core/utils/logging/log_service.dart';
 import 'package:spark/src/core/utils/logging/logger.dart';
+import 'package:spark/src/features/auth/providers/auth_providers.dart';
 import 'package:spark/src/features/stories/providers/story_auto_delete_provider.dart';
-import 'package:spark/src/features/stories/providers/story_provider_dependencies.dart';
+import 'package:spark/src/features/stories/providers/story_repository_provider.dart';
 
 part 'story_manager_provider.g.dart';
+
+final storyManagerLoggerProvider = Provider<SparkLogger>((ref) {
+  return GetIt.instance<LogService>().getLogger('StoryManager');
+});
 
 /// Simple state holder for the story manager
 class StoryManagerState {
@@ -34,25 +42,25 @@ class StoryManagerState {
 
 @riverpod
 class StoryManager extends _$StoryManager {
-  late final StoryRepository _repository;
-  late final SparkLogger _logger;
+  late final StoryRepository _repository = ref.read(storyRepositoryProvider);
+  late final SparkLogger _logger = ref.read(storyManagerLoggerProvider);
 
   @override
   Future<StoryManagerState> build() async {
-    _repository = ref.read(storyRepositoryProvider);
-    _logger = ref.read(storyLoggerProvider('StoryManager'));
+    ref.watch(currentDidProvider);
+    ref.watch(atprotoProvider);
     ref.read(storyAutoDeleteExecutorProvider.future).ignore();
     return _loadInitial();
   }
 
   Future<StoryManagerState> _loadInitial() async {
     try {
-      final did = ref.read(storyCurrentDidProvider);
+      final did = ref.read(currentDidProvider);
       if (did == null) {
         return StoryManagerState(stories: const [], error: 'Not authenticated');
       }
       // Page through all story records directly via atproto to include expired
-      if (!ref.read(storyAtprotoAvailableProvider)) {
+      if (ref.read(atprotoProvider) == null) {
         return StoryManagerState(
           stories: const [],
           error: 'AtProto not initialized',
