+++ lib/src/features/stories/providers/story_repository_provider.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:get_it/get_it.dart';
+import 'package:spark/src/core/network/atproto/data/repositories/story_repository.dart';
+
+final storyRepositoryProvider = Provider<StoryRepository>((ref) {
+  return GetIt.instance<StoryRepository>();
+});
